Put configs in `types` module
